feature: Make the name of a tile sticky
Before, the name of a tile was not visible if only part of the tile was on the screen. In this case, the name of the tile was off the screen. Now, the name of a tile has been made sticky so it always remains visible as long as part of the tile is visible as well.